requirements.txt
psycopg2-binary>=2.9
pandas>=1.5
numpy>=1.24
sqlalchemy>=1.4
python-dateutil